Extracted ButtonBar component from AddressBar
